Move all objects to the objects module
