+++ eitprocessing/loading/draeger.py
@@ -8,14 +8,14 @@ from typing import TYPE_CHECKING, NamedTuple
 
 import numpy as np
 
-from eitprocessing.continuousdata import ContinuousData
-from eitprocessing.datacollection import DataCollection
-from eitprocessing.eitdata import EITData, Vendor
-from eitprocessing.event import Event
 from eitprocessing.loading import load_eit_data
 from eitprocessing.loading.reader import Reader
-from eitprocessing.phases import MaxValue, MinValue
-from eitprocessing.sparsedata import SparseData
+from eitprocessing.objects.continuousdata import ContinuousData
+from eitprocessing.objects.datacollection import DataCollection
+from eitprocessing.objects.eitdata import EITData, Vendor
+from eitprocessing.objects.event import Event
+from eitprocessing.objects.phases import MaxValue, MinValue
+from eitprocessing.objects.sparsedata import SparseData
 
 if TYPE_CHECKING:
     from pathlib import Path
