+++ eitprocessing/objects/continuousdata.py
