+++ eitprocessing/objects/sparsedata.py
